+++ server/src/main/java/org/allaymc/server/network/processor/impl/ingame/MobEquipmentPacketProcessor.java
@@ -11,8 +11,7 @@ import org.cloudburstmc.protocol.bedrock.packet.MobEquipmentPacket;
 public class MobEquipmentPacketProcessor extends PacketProcessor<MobEquipmentPacket> {
     @Override
     public void handleSync(EntityPlayer player, MobEquipmentPacket packet, long receiveTime) {
-        var handSlot = packet.getHotbarSlot();
-        player.setHandSlot(handSlot);
+        player.setHandSlot(packet.getHotbarSlot(), false);
     }
 
     @Override