+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and any changes to API will have a prefix `(API)`.
 
 Unless otherwise specified, any version comparison below is the comparison of server version, not API version.
 
-## 0.4.2 (API 0.7.0) - Unreleased
+## 0.5.0 (API 0.8.0) - Unreleased
 
 <small>[Compare with 0.4.1](https://github.com/AllayMC/Allay/compare/0.4.1...HEAD)</small>
 
@@ -18,6 +18,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 
 - (API) Introduced new option `raknet-packet-limit` in `ServerSettings` to control the maximum number of datagram
   packets each client can send within a single RakNet tick.
+- Added docker and docker-compose support, now you can build docker image by your own (image will be uploaded to Docker Hub in the future).
 
 ### Changed